LayoutTests/fast/regex/malformed-escapes-expected.txt

This page tests handling of malformed escape sequences.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Testing regexp: /\ug/gm
PASS regexp.test('ug') is true
PASS regexp.lastIndex is 2

Testing regexp: /\xg/gm
PASS regexp.test('xg') is true
PASS regexp.lastIndex is 2

Testing regexp: /\c_/gm
PASS regexp.test('\\c_') is true
PASS regexp.lastIndex is 3

Testing regexp: /[\B]/gm
PASS regexp.test('B') is true
PASS regexp.lastIndex is 1

Testing regexp: /[\b]/gm
PASS regexp.test('\b') is true
PASS regexp.lastIndex is 1

Testing regexp: /\8/gm
PASS regexp.test('\\8') is true
PASS regexp.lastIndex is 2
PASS successfullyParsed is true

TEST COMPLETE
